+++ packages/keycloak/test/factory/index.ts
@@ -1,0 +1,3 @@
+export * from './user.factory'
+export * from './policy.factory'
+export * from './resource.factory'