+++ packages/keycloak/test/factory/policy.factory.ts
@@ -1,0 +1,24 @@
+import * as faker from 'faker'
+import axios from 'axios'
+
+export class PolicyFactory {
+  constructor(private readonly accessToken: string) {}
+
+  async create(userIds: string[]): Promise<unknown> {
+    const { data } = await axios.post(
+      `${process.env.KEYCLOAK_SERVER_URL}/auth/realms/skore/authz/protection/uma-policy/18ee88a4-cd72-4e0d-9304-15abe6101b45`,
+      {
+        name: faker.name.title(),
+        scopes: ['create'],
+        users: userIds,
+      },
+      {
+        headers: {
+          Authorization: `Bearer ${this.accessToken}`,
+        },
+      },
+    )
+
+    return data
+  }
+}