+++ packages/keycloak/test/service/find-user.service.test.ts
@@ -2,7 +2,7 @@ import { suite, test, timeout } from '@testdeck/jest'
 import { FindUserService } from '../../src/service'
 import { BaseTest } from '../base-test'
 
-@suite('[Keycloak Module] Fid User Service')
+@suite('[Keycloak Module] Find User Service')
 export class FindUserServiceTest extends BaseTest {
   @test(timeout(10000))
   async 'Given a valid access token then return the user info'() {
@@ -14,6 +14,7 @@ export class FindUserServiceTest extends BaseTest {
     expect(user.email).toBeUndefined()
     expect(user.name).toBeUndefined()
     expect(user.organizationId).toBeUndefined()
+    expect(user.accessToken).toEqual(super.commonUserAccessToken())
   }
 
   @test(timeout(10000))