+++ packages/mongo/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.18.0](https://github.com/skore-io/nestjs-extensions/compare/v0.17.0...v0.18.0) (2020-12-04)
+
+**Note:** Version bump only for package @skore-io/mongo
+
+
+
+
+
+## [0.17.0](https://github.com/skore-io/nestjs-extensions/compare/v0.16.0...v0.17.0) (2020-12-03)
+
+**Note:** Version bump only for package @skore-io/mongo
+
+
+
+
+
+## [0.16.0](https://github.com/skore-io/nestjs-extensions/compare/v0.15.1...v0.16.0) (2020-12-02)
+
+**Note:** Version bump only for package @skore-io/mongo
+
+
+
+
+
+## [0.15.0](https://github.com/skore-io/nestjs-extensions/compare/v0.14.0...v0.15.0) (2020-12-01)
+
+**Note:** Version bump only for package @skore-io/mongo
+
+
+
+
+
+## [0.14.0](https://github.com/skore-io/nestjs-extensions/compare/v0.13.2...v0.14.0) (2020-11-30)
+
+**Note:** Version bump only for package @skore-io/mongo
+
+
+
+
+
+## [0.13.0](https://github.com/skore-io/nestjs-extensions/compare/v0.12.0...v0.13.0) (2020-11-13)
+
+**Note:** Version bump only for package @skore-io/mongo
+
+
+
+
+
+## [0.12.0](https://github.com/skore-io/nestjs-extensions/compare/v0.11.0...v0.12.0) (2020-11-13)
+
+**Note:** Version bump only for package @skore-io/mongo
+
+
+
+
+
+## [0.11.0](https://github.com/skore-io/nestjs-extensions/compare/v0.10.0...v0.11.0) (2020-11-12)
+
+**Note:** Version bump only for package @skore-io/mongo
+
+
+
+
+
+## [0.10.0](https://github.com/skore-io/nestjs-extensions/compare/v0.9.0...v0.10.0) (2020-11-11)
+
+**Note:** Version bump only for package @skore-io/mongo
+
+
+
+
+
 ## [0.9.0](https://github.com/skore-io/nestjs-extensions/compare/v0.8.0...v0.9.0) (2020-11-10)
 
 **Note:** Version bump only for package @skore-io/mongo